+++ RecoLocalMuon/GEMRecHit/interface/GEMEtaPartitionMask.h
@@ -1,5 +1,5 @@
-#ifndef RecoLocalMuon_GEMEtaPartitionMask_h
-#define RecoLocalMuon_GEMEtaPartitionMask_h
+#ifndef RecoLocalMuon_GEMRecHit_GEMEtaPartitionMask_h
+#define RecoLocalMuon_GEMRecHit_GEMEtaPartitionMask_h
 
 #include <bitset>
 #include <vector>
