+++ RecoLocalMuon/GEMRecHit/plugins/GEMRecHitAlgoFactory.cc
