+++ RecoLocalMuon/GEMRecHit/plugins/GEMRecHitAlgoFactory.h
@@ -1,6 +1,5 @@
-#ifndef RecoLocalMuon_GEMRecHitAlgoFactory_H
-#define RecoLocalMuon_GEMRecHitAlgoFactory_H
-
+#ifndef RecoLocalMuon_GEMRecHit_GEMRecHitAlgoFactory_H
+#define RecoLocalMuon_GEMRecHit_GEMRecHitAlgoFactory_H
 /** \class GEMRecHitAlgoFactory
  *  Factory of seal plugins for 1D RecHit reconstruction algorithms.
  *  The plugins are concrete implementations of GEMRecHitBaseAlgo base class.
@@ -12,7 +11,3 @@
 
 typedef edmplugin::PluginFactory<GEMRecHitBaseAlgo *(const edm::ParameterSet &)> GEMRecHitAlgoFactory;
 #endif
-
-
-
-
