+++ RecoLocalMuon/GEMRecHit/plugins/GEMRecHitStandardAlgo.cc
@@ -3,33 +3,22 @@
  *
  *  \author M. Maggi -- INFN
  */
-
-#include "GEMCluster.h"
-#include "RecoLocalMuon/GEMRecHit/src/GEMRecHitStandardAlgo.h"
+#include "RecoLocalMuon/GEMRecHit/interface/GEMCluster.h"
+#include "RecoLocalMuon/GEMRecHit/interface/GEMRecHitStandardAlgo.h"
 #include "DataFormats/MuonDetId/interface/GEMDetId.h"
 #include "Geometry/GEMGeometry/interface/GEMEtaPartition.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "FWCore/Framework/interface/EventSetup.h"
 #include "FWCore/Utilities/interface/Exception.h"
 
-
 GEMRecHitStandardAlgo::GEMRecHitStandardAlgo(const edm::ParameterSet& config) :
   GEMRecHitBaseAlgo(config) 
-{
-}
-
-
+{}
 
 GEMRecHitStandardAlgo::~GEMRecHitStandardAlgo()
-{
-}
+{}
 
-
-
-void GEMRecHitStandardAlgo::setES(const edm::EventSetup& setup) {
-}
-
-
+void GEMRecHitStandardAlgo::setES(const edm::EventSetup& setup) {}
 
 // First Step
 bool GEMRecHitStandardAlgo::compute(const GEMEtaPartition& roll,
@@ -49,7 +38,6 @@ bool GEMRecHitStandardAlgo::compute(const GEMEtaPartition& roll,
   return true;
 }
 
-
 bool GEMRecHitStandardAlgo::compute(const GEMEtaPartition& roll,
 				    const GEMCluster& cl,
 				    const float& angle,
@@ -57,11 +45,9 @@ bool GEMRecHitStandardAlgo::compute(const GEMEtaPartition& roll,
 				    LocalPoint& Point,
 				    LocalError& error)  const
 {
-
   // Glob Pos and angle not used so far...
   if (globPos.z()<0){ } // Fake use to avoid warnings
   if (angle<0.){ }      // Fake use to avoid warnings
   this->compute(roll,cl,Point,error);
   return true;
 }
-
