+++ RecoLocalMuon/GEMRecHit/plugins/SealModule.cc
@@ -1,0 +1,19 @@
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "RecoLocalMuon/GEMRecHit/plugins/GEMRecHitAlgoFactory.h"
+#include "RecoLocalMuon/GEMRecHit/plugins/ME0RecHitAlgoFactory.h"
+
+#include "RecoLocalMuon/GEMRecHit/plugins/GEMRecHitProducer.h"
+DEFINE_FWK_MODULE(GEMRecHitProducer);
+
+#include "RecoLocalMuon/GEMRecHit/plugins/GEMRecHitStandardAlgo.h"
+DEFINE_EDM_PLUGIN (GEMRecHitAlgoFactory, GEMRecHitStandardAlgo, "GEMRecHitStandardAlgo");
+
+
+#include "RecoLocalMuon/GEMRecHit/plugins/ME0RecHitProducer.h"
+DEFINE_FWK_MODULE(ME0RecHitProducer);
+
+#include "RecoLocalMuon/GEMRecHit/plugins/ME0RecHitStandardAlgo.h"
+DEFINE_EDM_PLUGIN (ME0RecHitAlgoFactory, ME0RecHitStandardAlgo, "ME0RecHitStandardAlgo");
+
