+++ RecoLocalMuon/GEMRecHit/src/GEMCluster.cc
@@ -1,4 +1,4 @@
-#include "GEMCluster.h"
+#include "RecoLocalMuon/GEMRecHit/interface/GEMCluster.h"
 #include <iostream>
 #include <fstream>
