--- RecoLocalMuon/GEMRecHit/src/GEMClusterContainer.h
@@ -1,6 +1,0 @@
-#ifndef RecoLocalMuon_GEMClusterContainer_h
-#define RecoLocalMuon_GEMClusterContainer_h
-#include <set>
-class GEMCluster;
-typedef std::set<GEMCluster> GEMClusterContainer;
-#endif
